Reduce Kotlin/Gradle warnings and clean nullable/deprecation issues

+++ downloader/core/src/commonMain/kotlin/ir/amirab/downloader/part/MediaSegment.kt
@@ -31,7 +31,7 @@ data class MediaSegment(
         }
 
     override fun getID(): Long {
-        return segmentIndex.toLong()
+        return segmentIndex
     }
 
     companion object {
